[FIT2] Pass lego_ib_test!
The main thing is to disable CONFIG_VMAP_STACK.

If the VMAP_STACK is enabled, the kernel stack is virtually-mapped,
which is also the root cause of 0c37f55fefb5b046ac9cbfc404dccd34cedffbf6.
By simply disabling it, we don't need to use kmalloc which is more
comlicated modifications.

For more information about virtually-mapped stack, see
https://lkml.org/lkml/2016/6/15/1064.